+++ src/main/java/net/epoxide/elysian/world/WorldProviderElysian.java
@@ -50,7 +50,7 @@ public class WorldProviderElysian extends WorldProvider {
         
         for (int i = 0; i <= 15; ++i) {
             
-            float f1 = 1.0F - (float) i / 15.0F;
+            float f1 = 1.0F - i / 15.0F;
             this.lightBrightnessTable[i] = (1.0F - f1) / (f1 * 3.0F + 1.0F) * (1.0F - f) + f;
         }
     }
